Moved extension classes into their own namespace. Attention: public APIs have been moved!
